requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==1.4.49
python-dotenv==1.0.0
email-validator==2.0.0
Jinja2==3.1.2
itsdangerous==2.1.2
alembic==1.12.0 